ragheat_crewai/requirements.txt
# CrewAI Multi-Agent Portfolio Construction System Requirements

# Core CrewAI Framework - Latest Version
crewai>=0.186.1
crewai-tools>=0.71.0

# Language Models
langchain>=0.3.27
langchain-anthropic>=0.3.20
langchain-community>=0.3.29
langchain-core>=0.3.76
langchain-text-splitters>=0.3.11
anthropic>=0.67.0

# Data Processing & Analysis
pandas>=2.1.1
numpy>=1.24.3
scipy>=1.11.1
scikit-learn>=1.3.0

# Financial Data APIs
yfinance>=0.2.31
alpha-vantage>=2.3.1
finnhub-python>=2.4.20

# Technical Analysis
ta-lib>=0.4.26
pandas-ta>=0.3.14b0

# Sentiment Analysis & NLP
transformers>=4.40.2
torch>=2.0.1
nltk>=3.8.1
textblob>=0.17.1
vaderSentiment>=3.3.2

# Knowledge Graph & Database
neo4j>=5.19.0
rdflib>=7.0.0
networkx>=3.2
py2neo>=2021.2.4

# Portfolio Optimization
cvxpy>=1.4.3
pypfopt>=1.5.5

# Visualization
matplotlib>=3.7.2
seaborn>=0.12.2
plotly>=5.14.1

# Web Framework & APIs
fastapi>=0.116.1
uvicorn>=0.35.0
requests>=2.32.5
httpx>=0.28.1

# Configuration & Environment
pydantic>=2.11.9
python-dotenv>=1.0.0
pyyaml>=6.0.2

# Utilities
python-dateutil>=2.8.2
pytz>=2025.2
tqdm>=4.67.1

# Logging & Monitoring
loguru>=0.7.0

# Redis for caching
redis>=4.5.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Jupyter notebooks for research
jupyter>=1.0.0
ipykernel>=6.23.1

# Additional dependencies for production
schedule>=1.2.0
jsonschema>=4.25.1
click>=8.2.1